requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11
django-cors-headers==2.0.2
django-debug-toolbar==1.7
django-excel==0.0.6
djangorestframework==3.6.2
et-xmlfile==1.0.1
jdcal==1.3
openpyxl==2.4.6
packaging==16.8
pyexcel==0.4.5
pyexcel-io==0.3.3
pyexcel-webio==0.0.11
pyexcel-xls==0.3.1
pyexcel-xlsx==0.3.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
sqlparse==0.2.3
texttable==0.8.8
xlrd==1.0.0
xlwt==1.2.0
